Moved gui to its own file
